Added a flag for forced person creation
The 'force' flag has to be set to true, if there is already another person with the same name (yes, in churches with a larger number of members it may happen that there are different persons with the same name ;-)) and the person should be created anyways. Otherwise the API response with an error.