Sending API key as HTTP header instead of query param in FileRequestBuilder
Using the HTTP header for this scenario is a little bit cleaner as it is designed exactly for this use case.

Further more eventually existing trailing slashes in the API URL are trimmed to prevent a double slash in the URL like `https://example.church.tools//api/files/...` which may lead to issues.